refactor: remove 'popup' as this isn't helpful
