+++ .changeset/green-cats-speak.md
@@ -1,0 +1,5 @@
+---
+'@stacks/wallet-web': patch
+---
+
+Refactor of our component organisation, making efforts to structure by feature, and giving clearer responsibilities to each top level directory
