+++ src/app.tsx
@@ -1,17 +1,17 @@
 import React, { useEffect } from 'react';
-import packageJson from '../../package.json';
 import { ThemeProvider, ColorModeProvider } from '@stacks/ui';
+import { Toaster } from 'react-hot-toast';
 import { RecoilRoot } from 'recoil';
+import packageJson from '../package.json';
 
 import { theme } from '@common/theme';
-import { Routes } from '@components/routes';
 import { HashRouter as Router } from 'react-router-dom';
 import { GlobalStyles } from '@components/global-styles';
 import { VaultLoader } from '@components/vault-loader';
-import { AccountsDrawer } from './drawer/accounts';
-import { NetworksDrawer } from './drawer/networks-drawer';
-import { Toaster } from 'react-hot-toast';
+import { AccountsDrawer } from './components/drawer/accounts';
+import { NetworksDrawer } from './components/drawer/networks-drawer';
 import { DebugObserver } from '@components/debug-observer';
+import { Routes } from './routes';
 
 export const App: React.FC = () => {
   useEffect(() => {
