+++ src/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from '@components/app';
+import App from './app';
 
 const buildApp = () => {
   ReactDOM.render(<App />, document.getElementById('actions-root'));
