+++ src/pages/transaction-signing/components/attachment-row.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import { hexToHumanReadable } from '@common/utils';
-import { useTransactionRequest } from '../hooks/use-transaction';
+import { useTransactionRequest } from '../../../common/hooks/use-transaction-request';
 import { RowItem } from './row-item';
 
 export const AttachmentRow: React.FC = () => {
