+++ src/pages/transaction-signing/components/post-conditions/list.tsx
@@ -9,7 +9,7 @@ import { truncateMiddle } from '@stacks/ui-utils';
 import { useLoadable } from '@common/hooks/use-loadable';
 import { postConditionsState } from '@store/transactions';
 
-import { useTransactionRequest } from '../../hooks/use-transaction';
+import { useTransactionRequest } from '../../../../common/hooks/use-transaction-request';
 import { TransactionEventCard } from '../event-card';
 import { PostConditionComponent } from './single';
