+++ src/pages/transaction-signing/components/stx-transfer-details.tsx
@@ -3,7 +3,7 @@ import { color, Stack } from '@stacks/ui';
 import { AttachmentRow } from './attachment-row';
 import { Title } from '@components/typography';
 import { Divider } from '@components/divider';
-import { useTransactionRequest } from '@pages/transaction-signing/hooks/use-transaction';
+import { useTransactionRequest } from '@common/hooks/use-transaction-request';
 import { RowItem } from './row-item';
 
 export const StxTransferDetails: React.FC = () => {
