+++ src/pages/transaction-signing/hooks/use-transaction.ts
@@ -1,16 +1,10 @@
 import { useLoadable } from '@common/hooks/use-loadable';
-import { requestTokenPayloadState } from '@store/transactions/requests';
 import { postConditionsState, signedTransactionState } from '@store/transactions';
 import {
   transactionContractInterfaceState,
   transactionContractSourceState,
   transactionFunctionsState,
 } from '@store/transactions/contract-call';
-
-export function useTransactionRequest() {
-  const payload = useLoadable(requestTokenPayloadState);
-  return payload?.value;
-}
 
 export function useTransactionContractInterface() {
   return useLoadable(transactionContractInterfaceState);
